Refactor package and group ID to com.arran4.txtar
Updates the project configuration and package structure to use `com.arran4.txtar`.
This avoids hyphens in the package name (which are invalid) while aligning with the requested namespace.

- Updates `pluginGroup` in `gradle.properties` to `com.arran4.txtar`.
- Updates `group` in `build.gradle.kts`.
- Updates `plugin.xml` ID to `com.arran4.txtar` and vendor to `arran4`.
- Moves all Kotlin source files to `com.arran4.txtar`.
- Updates all package declarations and imports.
