+++ build.gradle.kts
@@ -4,7 +4,7 @@ plugins {
     id("org.jetbrains.intellij") version "1.17.3"
 }
 
-group = "com.github.arran4.idea.txtar"
+group = "com.arran4.txtar"
 version = System.getenv("GITHUB_REF_NAME") ?: "1.0.0"
 
 repositories {
