+++ gradle.properties
@@ -1,4 +1,4 @@
-pluginGroup=com.github.arran4.idea.txtar
+pluginGroup=com.arran4.txtar
 pluginName=txtar-support
 pluginVersion=1.0.0
 pluginSinceBuild=232
