+++ src/main/kotlin/com/arran4/txtar/TxtarLexer.kt
@@ -1,4 +1,4 @@
-package com.github.arran4.idea.txtar
+package com.arran4.txtar
 
 import com.intellij.lexer.LexerBase
 import com.intellij.psi.tree.IElementType
